TYPING SPEED TEST
The Typing Speed Test Project is a simple Java application designed to measure a user's typing speed in Words Per Minute (WPM). The program challenges the user to type a pre-defined sentence as quickly and accurately as possible. It calculates the number of words typed and the time taken, providing an accurate assessment of typing speed